google-cloud-function-src/requirements.txt
google-resumable-media
google-cloud-storage
google-cloud-bigquery
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cryptography==3.4.8
cycler==0.10.0
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
packaging==21.0
pandas==1.1.5
Pillow==8.3.2
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
redis==3.4.1
regex==2021.8.28
requests==2.24.0
rq==1.2.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.96
six==1.16.0
SQLAlchemy==1.4.23
threadpoolctl==2.2.0
tokenizers==0.8.1rc2
torch==1.5.1
torchvision==0.6.1
tqdm==4.62.2
transformers==3.3.1
urllib3==1.25.11
Werkzeug==2.0.1